+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Finely tuned the `loading view` using `async_sleep()`
 * migrated from `rusqlite` to `sqlx with sqlite`
 
+### Fixed
+* If an error occurs in `reqwest::get()` in `cat search`, `Loading` remains displayed.
+
 
 ## [0.1.2] (2026-01-04)
 ### Changed