+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.1.3] (2026-01-13)
 ### Added
 * crate: `async-sleep-aki`
 * `bicmid_id` column into `Cat` table
@@ -18,7 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * If an error occurs in `reqwest::get()` in `cat search`, `Loading` remains displayed.
-
 
 ## [0.1.2] (2026-01-04)
 ### Changed
@@ -46,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * first commit
 
-[Unreleased]: https://github.com/aki-akaguma/cat_tongue/compare/v0.1.2..HEAD
+[Unreleased]: https://github.com/aki-akaguma/cat_tongue/compare/v0.1.3..HEAD
+[0.1.3]: https://github.com/aki-akaguma/cat_tongue/compare/v0.1.2..v0.1.3
 [0.1.2]: https://github.com/aki-akaguma/cat_tongue/compare/v0.1.1..v0.1.2
 [0.1.1]: https://github.com/aki-akaguma/cat_tongue/compare/v0.1.0..v0.1.1
 [0.1.0]: https://github.com/aki-akaguma/cat_tongue/releases/tag/v0.1.0